g2poly: Add extended euclidean algorithm
Add the extended euclidean algorithm for getting values x and y
such that gcd(a, b) == a * x + b * y. This is useful to calculate
the inverse element in finite fields.

As a requirement, also add a division operator to G2Poly for getting
the quotient of polynomial division.

Also add G2PolyProd::to_poly and G2PolyProd::try_to_poly methods
that convert to G2Poly.