g2gen: Switch to full precomputed tables for Mul, Div
Use precomputed tables for Mul and Div implementation.
This allows for faster * and / implementation for the generated type

To keep the Mul::mul table small, the items are split into byte sized
bins. Since we can use the distributive law (a+b)*(c+d) =
a*c+a*d+b*c+b*d we can compute the multiplication table just inbetween
these byte bins. This means we have 256*256*n*n table entries, where n
is the base 256 log of the number of elements in the field, (i.e. the
max. number of bytes needed to represent a number).

To speed up compile times, we generate the array "string" by hand before
passing it to the compiler. The quote! macro is too slow otherwise.